moved token_info to public service
