+++ ssas/service/admin/api.go
@@ -3,8 +3,6 @@ package admin
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/dgrijalva/jwt-go"
-	"github.com/go-chi/render"
 	"io/ioutil"
 	"net/http"
 	"strings"
@@ -499,42 +497,6 @@ func getSystemIPs(w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		jsonError(w, http.StatusInternalServerError, "internal error")
 	}
-}
-
-func validateAndParseToken(w http.ResponseWriter, r *http.Request) {
-	trackingID := uuid.NewRandom().String()
-	event := ssas.Event{Op: "V2-Token-Info", TrackingID: trackingID, Help: "calling from admin.validateAndParseToken()"}
-	ssas.OperationCalled(event)
-
-	defer r.Body.Close()
-
-	var reqV map[string]string
-	if err := json.NewDecoder(r.Body).Decode(&reqV); err != nil {
-		jsonError(w, http.StatusBadRequest, "invalid request body")
-		return
-	}
-	tokenS := reqV["token"]
-	if tokenS == "" {
-		jsonError(w, http.StatusBadRequest, `missing "token" field in body`)
-		return
-	}
-	var response = make(map[string]interface{})
-
-	if err := tokenValidity(tokenS, "AccessToken"); err != nil {
-		ssas.Logger.Infof("token failed tokenValidity")
-		response["valid"] = false
-	} else {
-		claims := jwt.MapClaims{}
-		if _, _, err := new(jwt.Parser).ParseUnverified(tokenS, claims); err != nil {
-			ssas.Logger.Infof("could not unmarshal access token")
-			jsonError(w, http.StatusInternalServerError, "internal server error")
-			return
-		}
-		response["valid"] = true
-		response["claims"] = claims
-	}
-	w.Header().Set("Content-Type", "application/json")
-	render.JSON(w, r, response)
 }
 
 type IPAddressInput struct {
