--- ssas/service/admin/tokens.go
@@ -1,89 +1,0 @@
-package admin
-
-import (
-	"fmt"
-	"github.com/CMSgov/bcda-ssas-app/ssas"
-	"github.com/CMSgov/bcda-ssas-app/ssas/service"
-)
-
-func init() {
-}
-
-func empty(arr []string) bool {
-	empty := true
-	for _, item := range arr {
-		if item != "" {
-			empty = false
-			break
-		}
-	}
-	return empty
-}
-
-func tokenValidity(tokenString string, requiredTokenType string) error {
-	tknEvent := ssas.Event{Op: "tokenValidity"}
-	ssas.OperationStarted(tknEvent)
-	t, err := server.VerifyToken(tokenString)
-	if err != nil {
-		tknEvent.Help = err.Error()
-		ssas.OperationFailed(tknEvent)
-		return err
-	}
-
-	c := t.Claims.(*service.CommonClaims)
-
-	err = checkAllClaims(c, requiredTokenType)
-	if err != nil {
-		tknEvent.Help = err.Error()
-		ssas.OperationFailed(tknEvent)
-		return err
-	}
-
-	err = c.Valid()
-	if err != nil {
-		tknEvent.Help = err.Error()
-		ssas.OperationFailed(tknEvent)
-		return err
-	}
-
-	if service.TokenBlacklist.IsTokenBlacklisted(c.Id) {
-		err = fmt.Errorf("token has been revoked")
-		tknEvent.Help = err.Error()
-		ssas.OperationFailed(tknEvent)
-		return err
-	}
-
-	ssas.OperationSucceeded(tknEvent)
-	return nil
-}
-
-func checkAllClaims(claims *service.CommonClaims, requiredTokenType string) error {
-	if err := server.CheckRequiredClaims(claims, requiredTokenType); err != nil {
-		return err
-	}
-
-	if err := checkTokenClaims(claims); err != nil {
-		return err
-	}
-	return nil
-}
-
-func checkTokenClaims(claims *service.CommonClaims) error {
-	switch claims.TokenType {
-	case "MFAToken":
-		if claims.OktaID == "" {
-			return fmt.Errorf("MFA token must have OktaID claim")
-		}
-	case "RegistrationToken":
-		if empty(claims.GroupIDs) {
-			return fmt.Errorf("registration token must have GroupIDs claim")
-		}
-	case "AccessToken":
-		if claims.Data == "" {
-			return fmt.Errorf("access token must have Data claim")
-		}
-	default:
-		return fmt.Errorf("missing token type claim")
-	}
-	return nil
-}
