+++ ssas/service/public/router.go
@@ -2,12 +2,12 @@ package public
 
 import (
 	"fmt"
-	"os"
-	"time"
-	"github.com/go-chi/chi"
 	"github.com/CMSgov/bcda-ssas-app/ssas"
 	"github.com/CMSgov/bcda-ssas-app/ssas/constants"
 	"github.com/CMSgov/bcda-ssas-app/ssas/service"
+	"github.com/go-chi/chi"
+	"os"
+	"time"
 )
 
 var infoMap map[string][]string
@@ -48,6 +48,9 @@ func routes() *chi.Mux {
 	router.With(parseToken, requireRegTokenAuth, readGroupID).Post("/reset", ResetSecret)
 
 	//v2 Routes
+
 	router.Post("/v2/token", tokenV2)
+	router.Post("/v2/token_info", validateAndParseToken)
+
 	return router
 }
